- fix that

+++ base/unix/src/bootstrap.cpp
@@ -18,7 +18,7 @@
 	along with this program; if not, write to the Free Software
 	Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
 	
-	$Id: bootstrap.cpp,v 1.33 2000/09/28 12:11:00 ijr Exp $
+	$Id: bootstrap.cpp,v 1.34 2000/09/28 13:21:33 ijr Exp $
 ____________________________________________________________________________*/
 
 #include "config.h"
@@ -124,13 +124,12 @@ int main(int argc, char **argv)
         if (iCmdSem < 0)
         {
            printf("Cannot create/open a semaphore. Is SYS V IPC installed?\n");
-           exit(0);
         }
 
         // Check to see if the process that created that semaphore still
         // exists
         iProcess = semctl(iCmdSem, 0, GETVAL, unsem);
-        if (iProcess > 0 && !allow_mult)
+        if (iProcess > 0 && !allow_mult && iCmdSem > 0)
         {
             if (kill(iProcess, 0) >= 0)
             {
@@ -166,16 +165,19 @@ int main(int argc, char **argv)
             }
         }
 
-        // Set the current pid into the semaphore
-        unsem.val = getpid();
-        semctl(iCmdSem, 0, SETVAL, unsem);
-
-        // Create the shared memory segment
-        iCmdMem = shmget(tMemKey, iSharedMemSize, IPC_CREAT | 0660);
-        if (iCmdMem != -1)
+        if (iCmdSem > 0) 
         {
-            pCmdLine = (char *)shmat(iCmdMem, NULL, 0); 
-            pCmdLine[0] = 0;
+            // Set the current pid into the semaphore
+            unsem.val = getpid();
+            semctl(iCmdSem, 0, SETVAL, unsem);
+
+            // Create the shared memory segment
+            iCmdMem = shmget(tMemKey, iSharedMemSize, IPC_CREAT | 0660);
+            if (iCmdMem != -1)
+            {
+                pCmdLine = (char *)shmat(iCmdMem, NULL, 0); 
+                pCmdLine[0] = 0;
+            }
         }
     }
 #endif
@@ -285,7 +287,7 @@ int main(int argc, char **argv)
     }
 
 #ifdef HAVE_SEMCTL
-    if (!allow_mult) {
+    if (!allow_mult && iCmdSem > 0) {
         if (pCmdLine)
             shmdt(pCmdLine);
         unsem.val = 0;
